devtools/client/webconsole/new-console-output/test/components/test_console-api-call_trace.html

<!DOCTYPE HTML>
<html lang="en">
<head>
  <meta charset="utf8">
  <title>Test for ConsoleApiCall component</title>
  <script type="text/javascript" src="chrome://mochikit/content/tests/SimpleTest/SimpleTest.js"></script>
  <script type="application/javascript;version=1.8" src="head.js"></script>
  <!-- Any copyright is dedicated to the Public Domain.
     - http://creativecommons.org/publicdomain/zero/1.0/ -->
</head>
<body>
<p>Test for ConsoleApiCall component with console.trace</p>

<script type="text/javascript;version=1.8">
"use strict";
window.onload = Task.async(function* () {
  const { prepareMessage } = require("devtools/client/webconsole/new-console-output/utils/messages");
  const { ConsoleApiCall } = require("devtools/client/webconsole/new-console-output/components/message-types/console-api-call");

  const packet = yield getPacket("console.trace()", "consoleAPICall");
  const message = prepareMessage(packet);
  const rendered = renderComponent(ConsoleApiCall, {message});
  is(rendered.getAttribute("open"), "true",
    "ConsoleApiCall has the expected value for the open attribute");

  const queryPath = "div.message.cm-s-mozilla span span.message-flex-body " +
    "span.message-body.devtools-monospace";
  const messageBody = rendered.querySelectorAll(queryPath);
  is(messageBody[0].textContent, "console.trace():",
    "ConsoleApiCall outputs expected text");

  const stacktraceQueryPath = "div.message.cm-s-mozilla ul.stacktrace";
  is(rendered.querySelectorAll(stacktraceQueryPath).length, 1,
    "ConsoleApiCall outputs one stacktrace element");
  SimpleTest.finish();
});
</script>
</body>
</html>
